requirements.txt
pycryptodome==3.20.0
pandas==2.2.3
matplotlib==3.9.2
